added wordpressID and refactored
